ml/requirements.txt
yfinance==0.2.36
ccxt==4.2.0
pandas==2.2.0
numpy==1.26.4
arch==6.3.0
hmmlearn==0.3.0
gymnasium==0.29.1
stable-baselines3==2.3.0
torch==2.2.0
cvxpy==1.4.2
scipy==1.12.0
matplotlib==3.8.3
seaborn==0.13.2
plotly==5.18.0
web3==6.15.0
eth-abi==5.0.0
pyyaml==6.0.1
tqdm==4.66.2
scikit-learn==1.4.0
python-dotenv==1.0.1
